Moved `chatHistory` to extension.
